# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Personelim API",
    "version": "1.0.0",
    "description": "Node.js backend API for personnel management system with authentication, employee management, leaves, advances, timesheets, payrolls, and salary payments.",
    "contact": {
      "name": "API Support",
      "email": "support@personelim.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://personelim-be.vercel.app",
      "description": "Production server (Vercel)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/employees/contract-types": {
      "get": {
        "summary": "Get available contract types",
        "tags": ["Employees"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Contract types retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contract types retrieved successfully"
                    },
                    "contractTypes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Belirsiz Süreli", "Belirli Süreli", "Kısmi Süreli", "Çağrı Üzerine", "Deneme Süreli"]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employees/work-modes": {
      "get": {
        "summary": "Get available work modes",
        "tags": ["Employees"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Work modes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Work modes retrieved successfully"
                    },
                    "workModes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Tam Zamanlı", "Yarı Zamanlı", "Part-time", "Hibrit", "Uzaktan (Remote)"]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}